refactor: extract SessionLibraryDateFilter.swift from SessionLibraryQuery.swift
Splits the `SessionLibraryDateFilter` enum (the biggest type in the
query bundle, with title + `matches(_:calendar:)` helpers) out of
SessionLibraryQuery.swift into its own file. SessionLibraryQuery.swift
retains the remaining query types (SortOrder, DateRange, Query).

Byte-identical move of lines 3-33. Both files import Foundation.

SessionLibraryQuery.swift: 61 → 30 lines. Tests: 667.
